Correcting choke timeout
The timeout for repetitive searches in PubMed was written
under the assumption that the used time function from QDateTime
would use milliseconds. However, as the function operates on
seconds, the timeout eventually became about one week instead
of 10 seconds.

+++ src/networking/onlinesearch/onlinesearchpubmed.h
@@ -38,7 +38,7 @@ public:
     virtual QUrl homepage() const;
 
     static const int maxNumResults;
-    static const qint64 queryChokeTimeout;
+    static const uint queryChokeTimeout;
 
 public slots:
     void cancel();
@@ -54,7 +54,7 @@ private:
     class OnlineSearchPubMedPrivate;
     OnlineSearchPubMedPrivate *d;
 
-    static qint64 lastQueryEpoch;
+    static uint lastQueryEpoch;
 };
 
 #endif // KBIBTEX_ONLINESEARCH_PUBMED_H
